backend/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.13.2
amqp==5.1.1
annotated-types==0.6.0
anyio==4.2.0
APIFlask==2.1.3
apispec==6.6.1
async-timeout==4.0.3
attrs==22.2.0
billiard==3.6.4.0
blinker==1.7.0
cachetools==5.3.0
celery==5.2.7
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==42.0.5
ecdsa==0.19.0
exceptiongroup==1.2.0
Flask==2.2.3
Flask-Cors==4.0.0
Flask-HTTPAuth==4.8.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
flask-marshmallow==1.2.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.0.3
frozenlist==1.3.3
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.0
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
h2==2.6.2
hpack==3.0.0
http-ece==1.2.0
httpcore==1.0.2
httplib2==0.22.0
httpx==0.26.0
hyper==0.7.0
hyperframe==3.2.0
idna==3.4
importlib-metadata==6.3.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.21.3
multidict==6.0.4
oauth2client==4.1.3
openai==0.27.4
packaging==23.2
Pillow==8.4.0
prompt-toolkit==3.0.38
protobuf==4.22.3
psycopg2-binary==2.9.6
py-vapid==1.9.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.22
pycryptodomex==3.20.0
pydantic==2.5.3
pydantic_core==2.14.6
PyJWT==2.8.0
pyparsing==3.0.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-gnupg==0.5.2
pytz==2023.3
pywebpush==2.0.0
redis==5.0.4
replicate==0.22.0
requests==2.28.2
rsa==4.9
sentry-sdk==2.5.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.31
stripe==9.9.0
supervisor==4.2.5
tqdm==4.65.0
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
webargs==8.4.0
Werkzeug==2.2.3
yarl==1.8.2
zipp==3.15.0
